refactor: move statics out of KeyFocusListener
